Convert Long to Integer for quote ID

+++ src/main/java/com/example/underwritingaidepoc/repository/QuoteRepository.java
@@ -12,6 +12,6 @@ public interface QuoteRepository extends JpaRepository<Quote, Long> {
 
     List<Quote> findAllQuotes();
 
-    Quote findQuoteById(Long id);
+    Quote findQuoteById(Integer id);
 
 }
